\name{earthquake}

\docType{data}

\alias{earthquake}

\title{Mt St Helens earthquake data}

\description{
  This data set is a reduced version of the full data set
  in Scott (1992). It contains the first three variables. 
}

\usage{data(earthquake)}

\format{A matrix with 3 columns and 510 rows.  Each row corresponds to
  the measurements of an earthquake beneath the Mt St Helens volcano.
  The first column is the longitude (in degrees, where a negative number
  indicates west of the International Date Line), the second column is
  the latitude (in degrees, where a positive number indicates north of
  the Equator) and the third column is the depth (in km, where a
  negative number indicates below the Earth's surface).
}

\source{
  Scott, D.W. (1992). \emph{Multivariate Density Estimation: Theory, Practice
    and Visualization.} John Wiley \& Sons Inc., New York.
  
}


\keyword{datasets}
